# nodejs-server-server/api/swagger.yaml
---
swagger: "2.0"
info:
  description: "FluxNotes Treatment Options Rest Client"
  version: "1.0.0"
  title: "FluxNotes Treatment Options Rest Client"
host: "localhost:3001"
basePath: "/api"
schemes:
- "http"
paths:
  /treatmentOptions:
    get:
      summary: "Get treatment options given filters"
      description: "Return treatment options for the following filters"
      operationId: "findTreatmentOptionsByPatientStats"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "Disease"
        in: "query"
        description: "disease filter"
        required: true
        type: "string"
      - name: "Race"
        in: "query"
        description: "race filter"
        required: false
        type: "string"
      - name: "dxGrade"
        in: "query"
        description: "grade of disease filter"
        required: false
        type: "string"
      responses:
        200:
          description: "array of treatment options for given filters"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/TreatmentOptions"
          headers:
            Access-Control-Allow-Origin:
              type: "string"
              description: "Setting this header to `*` allows all origins. This is\
                \ handy for public REST APIs that don't require authentication. But,\
                \ according to the HTTP spec, browsers *WILL NOT* send cookies if\
                \ this header is `*`, regardless of what you set `Access-Control-Allow-Credentials`\
                \ to.\n"
              default: "*"
        400:
          description: "Invalid filter stats applied"
          headers:
            Access-Control-Allow-Origin:
              type: "string"
              description: "Setting this header to `*` allows all origins. This is\
                \ handy for public REST APIs that don't require authentication. But,\
                \ according to the HTTP spec, browsers *WILL NOT* send cookies if\
                \ this header is `*`, regardless of what you set `Access-Control-Allow-Credentials`\
                \ to.\n"
              default: "*"
        404:
          description: "no treatment options found for filters"
          headers:
            Access-Control-Allow-Origin:
              type: "string"
              description: "Setting this header to `*` allows all origins. This is\
                \ handy for public REST APIs that don't require authentication. But,\
                \ according to the HTTP spec, browsers *WILL NOT* send cookies if\
                \ this header is `*`, regardless of what you set `Access-Control-Allow-Credentials`\
                \ to.\n"
              default: "*"
      x-swagger-router-controller: "Default"
    options:
      operationId: "treatmentOptionsOPTIONS"
      parameters: []
      responses:
        200:
          description: "Successful Data Retrivial"
          headers:
            Access-Control-Allow-Origin:
              type: "string"
              description: "Setting this header to `*` allows all origins. This is\
                \ handy for public REST APIs that don't require authentication. But,\
                \ according to the HTTP spec, browsers *WILL NOT* send cookies if\
                \ this header is `*`, regardless of what you set `Access-Control-Allow-Credentials`\
                \ to.\n"
              default: "*"
        400:
          description: "Invalid Filter applied"
          headers:
            Access-Control-Allow-Origin:
              type: "string"
              description: "Setting this header to `*` allows all origins. This is\
                \ handy for public REST APIs that don't require authentication. But,\
                \ according to the HTTP spec, browsers *WILL NOT* send cookies if\
                \ this header is `*`, regardless of what you set `Access-Control-Allow-Credentials`\
                \ to.\n"
              default: "*"
        404:
          description: "Not found"
          headers:
            Access-Control-Allow-Origin:
              type: "string"
              description: "Setting this header to `*` allows all origins. This is\
                \ handy for public REST APIs that don't require authentication. But,\
                \ according to the HTTP spec, browsers *WILL NOT* send cookies if\
                \ this header is `*`, regardless of what you set `Access-Control-Allow-Credentials`\
                \ to.\n"
              default: "*"
      x-swagger-router-controller: "Default"
definitions:
  TreatmentOptions:
    type: "object"
    properties:
      disease:
        type: "string"
        description: "disease filter"
      race:
        type: "string"
        description: "race filter"
      isAlive:
        type: "string"
        description: "dead or alive"
      survivalMonths:
        type: "integer"
        format: "int32"
        description: "survival time in months"
      dxGrade:
        type: "string"
        description: "grade of the disease to filter"
      treatOptions:
        type: "string"
        description: "treatment option recieved"
    example:
      isAlive: "isAlive"
      dxGrade: "dxGrade"
      treatOptions: "treatOptions"
      disease: "disease"
      race: "race"
      survivalMonths: 0
externalDocs:
  description: "Find out more about Swagger"
  url: "http://swagger.io"